# Vectorize Guid equality.