@gfoidl 2022-03-20
For reference: https://github.com/dotnet/runtime/pull/53012#issuecomment-845327403