@tannergooding 2022-04-25
I should've caught that the tests were slightly older.

The removal of `using System.Numerics;` is a build break since `IComparisonOperators` was moved to this namespace.

I have a fix up here: https://github.com/dotnet/runtime/pull/68494