@AndyAyersMS 2022-04-28
Improvements seen on windows arm64: https://github.com/dotnet/perf-autofiling-issues/issues/4993